ApiServer/app/requirements.txt
requests
msal
pyjwt